// front/src/core/apis/backend/generated/api.ts
export const AppArch = {
	Win32: "Win32",
	Win64: "Win64",
	LinuxDeb: "LinuxDeb",
	LinuxRpm: "LinuxRpm",
	LinuxSnap: "LinuxSnap",
} as const;

export type AppArch = typeof AppArch[keyof typeof AppArch];